requirements.txt
spotipy==2.26.0
requests==2.32.3
musicxmatch_api==1.0.7
boto3==1.42.0
matplotlib==3.8.4
jinja2==3.1.6
Pillow==10.4.0
python-dotenv==1.2.0
numpy==1.26.4
